refact: excluindo componentização desnecessaria da tabela de users
