+++ .gitignore
@@ -12,7 +12,7 @@ dist
 dist-ssr
 *.local
 
-.env
+*.env
 
 # Editor directories and files
 .vscode/*
